Prevent AtlasLoot from breaking on non-existant loot table
When a loot table does not exists, and AtlasLoot tries to fetch it, the whole thing breaks. This change will prevent that and also writes out for the client on which table is missing, for example: "AtlasLoot Error! KaraAttumenHEROIC not listed in loot table registry, please report this message to the AtlasLoot forums at http://www.atlasloot.net/"