add new context menu entry
When the auction house is open and you open the context menu of an item in atlas, A new entry will show called auction house. This will give you options to search for that item or if its a crafting recipe then options to search for the crafted item and recipe if its a drop recipe.